web-admin: UI ready for allowing flexibility add/search Parent
1. UI ready for adding child with optional Searching existing Parent or Adding new Parent

2. Handle case Parent2 is null, we want when edit or review is keeping step 2 is skipped and next() is going to the last step. All in HandleNext().

+++ backend/src/controllers/web-admin/TeacherController.ts
@@ -3,7 +3,7 @@
 
 import { Request, Response } from "express";
 import * as TeacherService from "../../services/web-admin/teacherService";
-import { daycareLocationIds } from "../../services/authService";
+import { canAdminnAccess, daycareLocationIds } from "../../services/authService";
 
 
 // POST /api/teachers
@@ -94,9 +94,6 @@ export const updateTeacher = async (req: Request, res: Response) => {
   if (!daycareId) {
     return res.status(400).json({message: "Daycare missing from current Admin  profile"});
   }
-  // Get the list of location Id: 
-  const locationIds = await daycareLocationIds(daycareId);
-
 
   try {
     const id = req.params.id;
@@ -109,29 +106,34 @@ export const updateTeacher = async (req: Request, res: Response) => {
     if (!teacher.locationId) {
       return res.status(403).json({ message: "Teacher data is missing location" });
     }
-    
-    // Checking admin permission 
     const teacherLocation = teacher.locationId;
     if (!teacherLocation) {
       return res.status(403).json({ message: "Missing LocationId from teacher" });
     }
+    // These are Factored out helper function inside of AuthService: 
     // Case admin owner
-    if (locationId === '*') {
-      if ( req.user) {
-        const daycaredId = req.user.daycareId;
-        const locations = await daycareLocationIds(daycaredId);
-        if (!locations.includes(teacherLocation)) {
-          return res.status(403).json({ message: "Forbidden: cannot delete teacher from another location" });
-        }
-      }
-    } else {
-        // Check if teacher belongs to admin's location
-      if (teacherLocation !== locationId) {
-        return res.status(403).json({ message: "Forbidden: cannot update teacher from another location" });
-      }
-    }
-    // Proceed to update
+    // if (locationId === '*') {
+    //   if ( req.user) {
+    //     const daycaredId = req.user.daycareId;
+    //     const locations = await daycareLocationIds(daycaredId);
+    //     if (!locations.includes(teacherLocation)) {
+    //       return res.status(403).json({ message: "Forbidden: cannot delete teacher from another location" });
+    //     }
+    //   }
+    // } else {
+    //     // Check if teacher belongs to admin's location
+    //   if (teacherLocation !== locationId) {
+    //     return res.status(403).json({ message: "Forbidden: cannot update teacher from another location" });
+    //   }
+    // }
 
+    // Checking admin permission 
+    const hasAccess = await canAdminnAccess(locationId, daycareId, teacherLocation);
+    if (!hasAccess) {
+      return res.status(403).json({ message: "Forbidden: cannot access teacher from another location" });
+    }
+
+    // Proceed to update
     const updated = await TeacherService.updateTeacher(id, req.body);
     if (!updated) return res.status(404).json({ message: "Failed to update teacher" });
 
@@ -149,6 +151,12 @@ export const deleteTeacher = async (req: Request, res: Response) => {
   if (!locationId) {
     return res.status(400).json({message: "Location missing from current Admin profile"});
   }
+  
+  const daycareId = req.user?.daycareId;
+
+  if (!daycareId) {
+    return res.status(400).json({message: "Daycare missing from current Admin  profile"});
+  }
 
   try {
     // Extract id from URL params
@@ -164,25 +172,15 @@ export const deleteTeacher = async (req: Request, res: Response) => {
       return res.status(404).json({ message: "Teacher not found" });
     }
 
-    // Checking admin permission 
     const teacherLocation = teacher.locationId;
     if (!teacherLocation) {
       return res.status(403).json({ message: "Missing LocationId from teacher" });
     }
-    // Case admin owner
-    if (locationId === '*') {
-      if ( req.user) {
-        const daycaredId = req.user.daycareId;
-        const locations = await daycareLocationIds(daycaredId);
-        if (!locations.includes(teacherLocation)) {
-          return res.status(403).json({ message: "Forbidden: cannot delete teacher from another location" });
-        }
-      }
-    } else {
-        // Check if teacher belongs to admin's location
-      if (teacherLocation !== locationId) {
-        return res.status(403).json({ message: "Forbidden: cannot delete teacher from another location" });
-      }
+
+    // Checking admin permission 
+    const hasAccess = await canAdminnAccess(locationId, daycareId, teacherLocation);
+    if (!hasAccess) {
+      return res.status(403).json({ message: "Forbidden: cannot access teacher from another location" });
     }
     // Proceed to delete
     const ok = await TeacherService.deleteTeacher(id);
